chore: combine test files
