+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/MultiSetInlineMethodsTest.kt
@@ -1,7 +1,6 @@
 package xyz.lbres.kotlinutils.set.multiset
 
 import kotlin.test.Test
-import xyz.lbres.kotlinutils.set.multiset.inline.* // ktlint-disable no-wildcard-imports no-unused-imports
 
 internal class MultiSetInlineMethodsTest {
     @Test fun testMapToSet() = runMapToSetTests()
